+++ lib/blogo/constants.rb
@@ -1,0 +1,4 @@
+module Blogo
+  # Used in the gemspec and in the footer for default layout
+  HOME_PAGE_URL = "https://github.com/greyblake/blogo"
+end