+++ CHANGELOG.markdown
@@ -1,6 +1,7 @@
 # v0.0.10 2014-10-27
 
 * [feature] Generator blogo:install to install intializer (greyblake)
+* [changed] Add link "Powered by Blogo" for the default layout (greyblake)
 
 # v0.0.9 2014-06-08